add the ability to generate Xauthority files
Some programs need a Xauthority file present, which is not automatically
generated by Xvfb or Xephyr. This patch adds this ability with the
use_xauth parameter. If this parameter is True, PyVirtualDisplay will
generate a Xauthority file for the display and it will set the right
environment variables.

If xauth is not installed when use_xauth=True is given, a
pyvirtualdisplay.xauth.NotFoundError is raised.